requirements.txt
streamlit==1.28.0
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.24.3 